requirements.txt
websockets
aiortc
opencv-python
aiohttp
av # For video frame handling
pathlib
ujson
setuptools
python-dotenv
pyserial
pynmea2